+++ PcbDesignSimuModeling.Core/Models/Resources/MixedResource.cs
@@ -8,5 +8,5 @@ public abstract class MixedResource : IResource
     public abstract double ResValueForProc(int requestId);
     public abstract void FreeResource(int requestId);
     public abstract IResource Clone();
-    public abstract double Cost { get; }
+    public abstract decimal Cost { get; }
 }
