+++ PcbDesignSimuModeling.Core/Models/Resources/Server/Server.cs
@@ -23,7 +23,7 @@ public class Server : SharedResource, INotifyPropertyChanged
 
     public override IResource Clone() => new Server(InternetSpeed) {ResValueConvolution = ResValueConvolution};
 
-    public override double Cost => Math.Round(
+    public override decimal Cost => (decimal) Math.Round(
         Math.Exp(1.0 + 125.0 / (InternetSpeed + 31.5)) * InternetSpeed * 7.5) - 4000;
 
     public event PropertyChangedEventHandler? PropertyChanged;
