+++ PcbDesignSimuModeling.Core/Models/Resources/UndividedResource.cs
@@ -8,5 +8,5 @@ public abstract class UndividedResource : IResource
     public abstract void FreeResource(int requestId);
 
     public abstract IResource Clone();
-    public abstract double Cost { get; }
+    public abstract decimal Cost { get; }
 }
